requirements.txt
absl-py
dominate
h5py
imgaug
numpy
opencv-python
pillow
protobuf
pyyaml
scikit-image
scipy
torch==1.4.0
torchvision==0.5.0
tqdm
SimpleITK

